51번 이슈 관련 PR요청

+++ JimCarry/src/main/java/com/study/jimcarry/api/ReqQuotationRequest.java
@@ -61,7 +61,7 @@ public class ReqQuotationRequest {
 	@JsonProperty("houseSize") //클라이언트에서 requestBody에 json으로 보낼 때 매핑
 	@Schema(name="houseSize", description="집 평수") //Swagger/OpenAPI 문서를 자동 생성
 	@NotNull(message="집 평수는 필수입니다.")
-    private BigDecimal houseSize;
+    private int houseSize;
 
     // 엘리베이터 여부
 	@JsonProperty("hasElevator") //클라이언트에서 requestBody에 json으로 보낼 때 매핑
