+++ JimCarry/src/main/java/com/study/jimcarry/model/UpdateReqQuotationDTO.java
@@ -43,7 +43,7 @@ public class UpdateReqQuotationDTO {
 
 	// 집 평수
 	@Schema(name = "houseSize", description = "집 평수")
-	private BigDecimal houseSize;
+	private int houseSize;
 
 	// 엘리베이터 여부
 	@Schema(name = "hasElevator", description = "엘리베이터 여부")
